Eliminate redundancy in config parsing
The data structure returned from parsing config.yaml is now used
internally to store all the driver script settings. Previously a layer
of additional parsing in read_files.py was used, which is no longer
necessary.

Config variables of type int or bool are preserved as such in the config
data structure, but are converted to strings by util/setenv() when
they're written to environment variables.