Redefine names for paths
Reduce the number of paths we have to keep track of (ie set as
environment variabes) to 5 nonredundant ones: locations of the code,
model and observational data, the working directory and the output
directory. The currently existing values assumed for all of these are
set as defaults in the argument parser.