Fix convective_transition_diag env vars
convective_transition_diag was setting environment variables in code and
using those to determine the filename of output plots. This created
problems when the filename-dependent boilerplate code was moved from the
POD into file_io. The hacky workaround is to define
convective_transition_diag's POD-specific env vars in its settings.yml
file, and to check for those in file_io/make_pod_html(). We'll need to
revisit this with a more permanent solution.