Add option to set PODs by model
Add optional 'pod_list' setting to the model specification in case_list
in config.yml, to give a list of PODs to run for that model only which
will override the default.

Note that we haven't implemented running on multiple models yet, the
code only runs on the first model in case_list and ignores the rest.