Specify translation through 'variable_convention' field
Add 'variable_convention' (internally mapped to the 'convention'
attribute) to Diagnostic(), DataManager() and util.VariableTranslator().
This is so we can specify the naming convention independently of the
model, and have PODs that use different conventions than the model.
DataManager._setup_pods() now translates each variable in the POD's
varlist from the POD's convention (to CF, stored in 'CF_name') and from
that to the model's convention (stored in 'name_in_model').