Implement PathManager singleton
This commit just adds the implementation and tests, it doesn't actually
use it in the rest of the code base.

Keep track of paths in singleton rather than environment variables.
PathManager has two methods, modelPaths and podPaths, which return a
dict with paths specific to each model and pod. These then go on to be
set as attributes of the DataManagerr or Diagnostic objects in the
relevant setup methods.