Rewrite GFDL /pp/ DataManager's _query_data()
Use new functionality from rewrite of DataManager. Rewrite directory
walking code (filtered_os_walk). Keep track of undertermined components
in a ComponentKey namedtuple.

We walk the /pp/ directory once. For each leaf directory, if .nc files
are found for a variable and date range we're searching for, we store
the undetermined components in data_files and a reference to the file in
_component_map. Files are populated later, in plan_data_fetch_hook (next
commit).