Set data level axis from "lev_coord" env var
Sub-PODs were failing to run due to a missing "xyz_dimensions.txt" file
for the model. This is generated by COMPOSITE/data_routine.ncl, which
looks for a pressure axis named "lev" or "plev". We change this to look
for an axis name starting with "lev" or "plev". In the former case,
units are assumed to be Pa, otherwise hPa.

This is a temporary hack: a valid solution will compare the reported
units of the Z axes directly.