Resolve all paths in PathManager
Define CLI PathAction which stores a list of settings that should be
resolved as paths. NOTE that any such setting needs to have
'"action":"PathAction" in its definition in the JSON file.

Move resolve_paths() from util.py to a method of PathManager. Make
PathManager a Namespace because it will be accessed through
ConfigManager. Have model_paths() and pod_paths() return Namespaces
instead of dicts.

NOTE: references to PathManager in rest of code haven't been updated.