Reword start_install instructions
Broken .rst :numref:s were created by referring to anchors corresponding
to third-level headers (^^^), which aren't assigned a section number in
the current configuration. Fix this by deleting or reassigning
references to third-level headers.

Move details on the mechanism for assigning conda environments to PODs
to the developer information section (dev_instruct.rst), as this isn't
needed for users to operate the framework.