Add placeholder datelabels for static files
Define FXDateRange, FXDateMin, FXDateMax to populate date range
information for static (time-independent) datasets. Currently date
information is set to None in these cases, which doesn't distinguish
between this and uninitalized values.

The FXDate* placeholders should raise FXDateException if you attempt to
treat them like real DateRanges, with the exception of
"contains"/"overlaps" methods. These are used by DataManager to
determine which files include the analysis time period, and will always
return True if one or both of the ranges are a FXDateRange.

The implementation of the classes is extremely hacky, and involves
adding special cases to the parent class logic. This should be cleaned
up in the future.