Initial commit of data model classes
Define a class hierarchy loosely following the CF conventions. These
classes will be refined by the POD's varlist entries, and then
further refined by the DataSource (to be added in a subsequent commit)
which will use them to generate the data query.

Add abstract classes to datelabel for the purposes of establishing
class membership only.