Fix how Diagnostic sets time coord
NOAA-GFDL/MDTF-diagnostics#40

In DataManager, need to initialize PODs before knowing full details
about the data source. In DataSet, we want to treat a collection of
variables at different frequencies as having the "same" time axis if
their date ranges are the same.

Define a DMGenericTimeCoordinate to represent this
"general"/"pre-specified" time coordinate, and add a .change_coord()
method that allows replacing fields on a coordinate as well as changing
their classes.