Implement request reordering by preprocessor
NOAA-GFDL/MDTF-diagnostics#40

Add an edit_request method to preprocessor functions, called after
variables are configured with DataManager-specific values. This has the
task of rewriting the linked graph of VarlistEntries and their
alternates to reflect the transformations that the preprocessor can do:
for example, the level extractor inserts the full 4D data as an
alternate for a 3D slice.