Refactor Preprocessor
Preprocessor instances are now associated with (an attribute of)
Diagnostics, rather than VarlistEntries, because they need to consider
the POD's varlist as a whole during edit_request(). All processing
functions now take an instance of the VarlistEntry being processed, as
well as the xaray dataset containing the data for that variable.