Move existing files into subpackages
This commit only renames files with 'git mv'; no file contents are
altered (so as not to confuse git). Move utils to a subpackage of /src;
include tests with the packages.

Move site-specific files to /sites; move configuration to src/data/;
move code meant to be shared with PODs in /shared.