Move exception defs to src.util.exceptions
NOAA-GFDL/MDTF-diagnostics#43

This cuts down on the bookkeeping for module imports, and allows the
same excpetions to be reused more effectively by child classes in
differnt modules. A Diagnostic-related exception may be raised by an
instance of that class, but the code for handling it may be in a
different module, which then has to remember to import it.