Reorder CLI; add data_manager as a plugin & remove preprocessor
Continue to use "data_manager" to refer to data source, for backwards
compatibility. Eliminate preprocessor choice from the CLI, as it makes
more sense to let it be dictated by the data_manager. Commit rough-draft
plugin CLIs for local file and CMIP6 data_managers.

NOTE: docs on CLI are not updated.