Implement experiment selection logic
Make experiment_key a tuple of keys for experiment attributes that must
be the same across all PODs, within a POD, and for each variable
(necessary for CMIP6 experiment selection logic). Select keys at each of
the three levels incrementally.

Ad an _id attribute to VarlistEntry, Diagnostic and DataSourceBase to
facilitate lookups.