Replace settings file in sites/local with README
Existing docs call for users to directly edit src/default_tests.jsonc,
but changes there would be overriden by the settings file in
sites/local. This will be the intended method in the next release, but
for now replace the config in sites/local with a placeholder text file
explaining the situation.