Instantiate FieldlistEntry with placeholder coordinates
Need to accommodate use case where a model naming convention defines
multiple coordinates of a given axis type (eg a parametric and
non-parametric Z axis). Define new DMPlacholder*Coordinate classes in
data_model which serve as placeholders for the axes_set members (ie they
indicate that a FieldlistEntry depends on a Z axis, but the specific one
is to be determined later, during translate()).

FieldlistEntries still indexed by (standard_name, axes_set) in
VariableTranslator, which won't cover all cases and will need to be
reimplemented in the future.