Improve error propagation in POD/VarlistEntry setup
Previously, an exception in DataSourceBase.setup_var() would deactivate
that variable's POD entirely. Change this so that it only deactivates
the var being processed; the var will be skipped by POD's query and
alternate variable iteration logic, and the POD will only be deactivated
if no alternates can be found for the deactivated var.