Robust logic for precip flux <-> rate conversion
Previous unit conversion inserted a factor of liquid water density
whenver possible (in terms of unit equivalence), which was asking for
trouble as more variables are added. Re-implement this logic properly in
the form of a preprocessor function (PrecipRateToFluxFunction) that only
applies the transformation if the variable has the correct standard
name.

Change handling of VariableTranslator failures: not a fatal error for
POD, since all it means is that variable not defined in the model
convention -- becomes a fatal error only if query runs out of alternate
variables.

Also refactor common logic needed in PrecipRateToFluxFunction,
ExtractLevelFunction to module-level functions in preprocessor.py, to
hopefully make writing edit_request() methods easier.