Fail gracefully when set_experiment() fails for a POD
Previously, if set_experiment() failed on a POD, an exception would be
raised and not caught until request_data(), which would result in all
PODs failing. Instead, when this happens, log the exception in the
problematic POD's ExceptionQueue, update the list of active PODs and
variables, and try to come up with an experiment attribute assignment
for the remaining active PODs. To accomplish this, split off
set_experiment into its own pseudo-while-loop, select_data(), which is
sandwiched in between query_data() and fetch_data().