Add CLI flag to ignore GFDL /pp/ model component
Define new command-line flag --any_components for the "GFDL_pp" and
"GFDL_auto" DataManagers. Default (and preexisting) behavior is to
require all variable for a POD to come from the same model component and
error out if that's not possible, but passing this flag will disable
that check. The DataManager will still attempt to select all variables
from the same component if possible.