Better verification of conda envs
Preexisting code for verifying that a selected conda environment exists
(prior to activation) would raise a non-fatal error if environments were
installed in a prefix directory (not in the system conda). Fix this by
simply looking for the environment's directory, rather than trying to
parse the output of `conda env list`.