Allow relative paths in wrapper script
Alter top-level "mdtf" wrapper script created by conda_env_setup.sh so
that relative paths passed as arguments will be interpreted relative to
the current working directory, instead of the MDTF repo directory.

This is done in a hacky way, by setting PYTHONPATH, which is needed
because the package is currently called as a module and not a script.
This will be fixed in the next release.