Restore ENSO_MSE to reduce size of diff
The ENSO_MSE POD is present but nonfunctional in /develop, and has been
removed for this reason in feature/v3alpha3 (once debugged, it will be
added to the code as PR NOAA-GFDL/MDTF-diagnostics#110). To simplify the
diff of feature/v3alpha3-pods-only against /develop, restore the
ENSO_MSE files as they currently exist in /develop.