Restrict "scalar coordinates" to vertical coord
"Scalar coordinate" syntax is what the package uses to specify slices of
arrays (eg. pressure levels). Modify the xarray accessor to only
identify scalar coordinates in the vertical axis, as this is currently
the only use case for this feature. Other logic (in the preprocessor and
the data model) doesn't make this assumption, so the ability to deal
with more general slicings can be re-implemented in the future.

This change was prompted by errors encountered in running the package on
single-column model output, which has X and Y axes of size 1.