Refactor DataSource inheritance
Have OnTheFlyDirectoryHierarchyQueryMixin inherit from an abstract
OnTheFlyFilesystemQueryMixin, which generalizes the catalog construction
process in an abstract method.

Move more user input validation logic into
data_manager.DataSourceAttributesBase, in particular setting
CASE_ROOT_DIR and validating the variable naming convention. The
convention is now a field on the DataSourceAttributes class, not an
attribute on the DataSource itself.