Initial commit of per-object logging
Define a second, or parallel, logging hierarchy that runs from
MDTFFramework -> DataSourceBase (cases) -> Diagnostic (pods) ->
VarlistEntry (variables), to implement a history of actions taken on
each object. The goal is that the case-level log will continue to be a
"standard" log, recording everything happening in the case in
chronological order, but the pod-level log will be organized into
pod-specific and variable-specific events.

Objects in the hierarchy now inherit from
core.MDTFObjectBase, which initializes the object's logger. The logger
is configured by MDTFObjectLoggerMixin for pods and variables, and
MDTFCaseLoggerMixin (which uses the logging config file) for cases. The
framework continues to log to the module-level (root) logger.