Reduce duplicate logs
Define a "*Event" class of non-fatal exceptions for use in the data
query/fetch/preprocess stage, and put in logic to log these at the INFO
level, rather than the ERROR level. This addresses
NOAA-GFDL/MDTF-diagnostics#99.

Add logic to only log events as error when a POD or case is deactivated.