Combine CLI options to --overwrite_file_metadata
Default behavior is to update missing file metadata with our record,
while logging a warning (not raising an error anymore). Setting this
flag disables the (default) reverse behavior, where our record is always
updated to accurately reflect the metadata in the file. Setting this
flag will cause the preprocessor to assume all file metadata is
untrustworthy: existing attributes won't be erased, but they'll be
overwritten in the event of a conflict with our record.