Always use mamba for conda env installs (#166)
* Always use mamba for conda env installs

Modify conda_env_setup.sh script to always use mamba when invoked. If
the mamba executable isn't on the $PATH when invoked, the installation
is bootstrapped by first creating a temporary conda environment named
"_MDTF_install_temp" and installing only mamba in it. If this is done,
the environment is deleted after the installs.

* Fix installation on linux

Co-authored-by: tsjackson-noaa <thomas.jackson@noaa.gov>