Add var name specification and search to ExplictFileDataSource
Add a 'var_name' attribute to the ExplictFileDataSource input
configuration file to allow the user to manually specify the variable's
translated name, i.e. the name of the variable in the data files the
user is providing.

Add guess_dependent_var() method to DefaultDatasetParser to guess the
name of the dependent variable when the translated name isn't found in
the data file. This is disabled for all cases except
ExplictFileDataSource, but we add the method here rather than to
MetadataRewriteParser because we suspect we'll need to use it in other
cases in the future.

Documentation for these options is updated.