Undo addition of 'attrs' dict to data model
Roll back this change made in ea40d3632a37355ed89ea7f1be853dbce59753b1
(although not the rest of that commit). Representing metadata with an
'attrs' dict gets the data model closer to xarray, but we really want
xarray's attrs to be more strucutred, and the property definitions
conflicted with dataclass fields (and introduced redundancy and
possibility for error when they worked right).