Propagate mdtf_dataclass changes to regex_dataclass
Changes made to mdtf_dataclass in previous commit break comopsition in
regex_dataclass: a __post_init__ method is always defined, in order to
do type coercion.

Fix this by making the regex_dataclass decorator
independent of the mdtf_dataclass decorator, and removing redundant
applications of both decorators in the code.