Update FRE wrapper (mtdf_gfdl.csh)
Update the FRE wrapper script to use current flags.

Do not load a python env module; instead run the same commands as the
interactive wrapper script to invoke the site-installed conda env before
running the package.

Disable functionality to make a copy of output for hosting on an
internal website, since updates aren't being made to Dora. Retain code
for when this functionality is re-enabled.