Convention defaults for GFDL_PP, LocalFile data sources
Set the default --convention assumed for SampleLocalFileDataSource to
"CMIP", to be consistent with what's written in the documentation.

Set the default --convention assumed for the GFDL_PP data source to
"GFDL", since this is the most common use case (but not the only one).
Mention this in the docs. Take logic to set this default value out of
the mdtf_gfdl.csh wrapper script.