Restore model component-specific options
Allow --component and --chunk_freq flags to be passed to the GFDL_PP
data source to restrict the data query to files with those attributes.

Add functionality to the FRE wrapper script to handle the real use
cases. Pass --multi_component to the wrapper to invoke "frepp mode"
(multiple runs of the package on same data, as it's being output from
postprocessing); pass --component_only to restrict the model component
used.

Update and expand the docs for these settings to explain all this.