Refine model component-specific logic in FRE flags
Remove the --multi_component/--any_component flag from the mdtf_gfdl.csh
wrapper, and replace it with --run_once.

This is because the sensible default behavior for the wrapper hould be
the opposite of what was previously implemented: the default scenario
for when the package is called from FRE is the incremental/online
processing use case ("scenario 3B" in the docs).

Non-default use cases are when the package is invoked only once, after
all needed data is postprocessed (scenario 3A), or when the user wants
to manually restrict the operation to the given model components. These
are invokes via the --run_once or --component_only flags in the wrapper
script.