Fix FRE integration and improve documentation (#245)
* Update FRE wrapper (mtdf_gfdl.csh)

Update the FRE wrapper script to use current flags.

Do not load a python env module; instead run the same commands as the
interactive wrapper script to invoke the site-installed conda env before
running the package.

Disable functionality to make a copy of output for hosting on an
internal website, since updates aren't being made to Dora. Retain code
for when this functionality is re-enabled.

* More debugging statements in FRE wrapper

* Pass through args in wrapper; set --convention default to GFDL

* Add more detail on wrapper script to GFDL docs

* Convention defaults for GFDL_PP, LocalFile data sources

Set the default --convention assumed for SampleLocalFileDataSource to
"CMIP", to be consistent with what's written in the documentation.

Set the default --convention assumed for the GFDL_PP data source to
"GFDL", since this is the most common use case (but not the only one).
Mention this in the docs. Take logic to set this default value out of
the mdtf_gfdl.csh wrapper script.

* Better regex in FRE wrapper

* Restore model component-specific options

Allow --component and --chunk_freq flags to be passed to the GFDL_PP
data source to restrict the data query to files with those attributes.

Add functionality to the FRE wrapper script to handle the real use
cases. Pass --multi_component to the wrapper to invoke "frepp mode"
(multiple runs of the package on same data, as it's being output from
postprocessing); pass --component_only to restrict the model component
used.

Update and expand the docs for these settings to explain all this.

* Refine model component-specific logic in FRE flags

Remove the --multi_component/--any_component flag from the mdtf_gfdl.csh
wrapper, and replace it with --run_once.

This is because the sensible default behavior for the wrapper hould be
the opposite of what was previously implemented: the default scenario
for when the package is called from FRE is the incremental/online
processing use case ("scenario 3B" in the docs).

Non-default use cases are when the package is invoked only once, after
all needed data is postprocessed (scenario 3A), or when the user wants
to manually restrict the operation to the given model components. These
are invokes via the --run_once or --component_only flags in the wrapper
script.

* Update all references to GFDL site installation

Remove all paths and references to the old GFDL site installation in
oar.gfdl.mdteam and replace with their counterparts in the oar.gfdl.mdtf
role account.

* Fix .rst formatting bug in docs

* More .rst tweaks

* Tweaks to default data source convention specification

Co-authored-by: Jess <20195932+wrongkindofdoctor@users.noreply.github.com>